terminando primera entrega

+++ BD/CRUD.js
@@ -1,4 +1,5 @@
 import {auth,db} from './conf'
+import swal from 'sweetalert';
 export function addBD(id,nombreBD,valores){
 
     auth.onAuthStateChanged(async user=>{
@@ -6,10 +7,11 @@ export function addBD(id,nombreBD,valores){
             try{
                 if (id === ""){
                    await db.collection('Usuario').doc(user.uid).collection(nombreBD).doc().set(valores)
+                   swal("Buen Trabajo", "Se ha agregado correctamente", "success");
                 }
                 else{
                     await db.collection('Usuario').doc(user.uid).collection(nombreBD).doc(id).update(valores)
-                    
+                    swal("Buen Trabajo", "Se ha actualizado correctamente", "success");   
                 }
             }
 
@@ -21,18 +23,30 @@ export function addBD(id,nombreBD,valores){
     })
 }
 export function deleteBD(nombreBD,id,valores){
-    auth.onAuthStateChanged(async user=>{
-        if (user!=null){
-            try{
-                await db.collection('Usuario').doc(user.uid).collection(nombreBD).doc(id).delete()
-                await db.collection('Usuario').doc(user.uid).collection(`Inactivo-${nombreBD}`).doc().set(valores)
-
-            }
-            catch (error){
-                console.log(error)
-
-            }
+    swal({
+        title: "Estas seguro que lo desea eliminar?",
+        text: "Si eliminas este dato queedara borrado permanentemente",
+        icon: "warning",
+        buttons: true,
+        dangerMode: true,
+      })
+      .then((willDelete) => {
+        if (willDelete) {
+            auth.onAuthStateChanged(async user=>{
+                if (user!=null){
+                    try{
+                        await db.collection('Usuario').doc(user.uid).collection(nombreBD).doc(id).delete()
+                        await db.collection('Usuario').doc(user.uid).collection(`Inactivo-${nombreBD}`).doc().set(valores)
+                        swal("Buen Trabajo", "Se ha eliminado correctamente", "success");
+                    }
+                    catch (error){
+                        console.log(error)
+        
+                    }
+                }
+            })
         }
-    })
+      });
+
 }
